feat(test): Add receipts to blockchain_tests to help debug root mismatches (#5)
